+++ smartcontract/programs/doublezero-serviceability/src/processors/resource/create.rs
@@ -1,0 +1,79 @@
+use crate::{error::DoubleZeroError, globalstate::globalstate_get};
+use borsh::BorshSerialize;
+use borsh_incremental::BorshDeserializeIncremental;
+#[cfg(test)]
+use solana_program::msg;
+use solana_program::{
+    account_info::{next_account_info, AccountInfo},
+    entrypoint::ProgramResult,
+    pubkey::Pubkey,
+};
+use std::fmt;
+
+#[derive(BorshSerialize, BorshDeserializeIncremental, PartialEq, Clone, Default)]
+pub struct ResourceCreateArgs {
+    pub resource_type: crate::resource::ResourceType,
+}
+
+impl fmt::Debug for ResourceCreateArgs {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(
+            f,
+            "ResourceCreateArgs {{ resource_type: {:?} }}",
+            self.resource_type
+        )
+    }
+}
+
+pub fn process_create_resource(
+    program_id: &Pubkey,
+    accounts: &[AccountInfo],
+    value: &ResourceCreateArgs,
+) -> ProgramResult {
+    let accounts_iter = &mut accounts.iter();
+
+    let resource_account = next_account_info(accounts_iter)?;
+    let associated_account = next_account_info(accounts_iter)?;
+    let globalstate_account = next_account_info(accounts_iter)?;
+    let globalconfig_account = next_account_info(accounts_iter)?;
+    let payer_account = next_account_info(accounts_iter)?;
+    let system_program = next_account_info(accounts_iter)?;
+
+    #[cfg(test)]
+    msg!("process_create_resource({:?})", value);
+
+    // Check if the payer is a signer
+    assert!(payer_account.is_signer, "Payer must be a signer");
+
+    // Check the owner of the accounts
+    assert_eq!(
+        globalstate_account.owner, program_id,
+        "Invalid GlobalState Account Owner"
+    );
+    assert_eq!(
+        globalconfig_account.owner, program_id,
+        "Invalid GlobalConfig Account Owner"
+    );
+    assert_eq!(
+        *system_program.unsigned_key(),
+        solana_program::system_program::id(),
+        "Invalid System Program Account Owner"
+    );
+
+    let globalstate = globalstate_get(globalstate_account)?;
+    if !globalstate.foundation_allowlist.contains(payer_account.key) {
+        return Err(DoubleZeroError::NotAllowed.into());
+    }
+
+    super::create_resource(
+        program_id,
+        resource_account,
+        associated_account,
+        globalconfig_account,
+        payer_account,
+        accounts,
+        value.resource_type,
+    )?;
+
+    Ok(())
+}